refactor: rename compliance dashboard to dashboard and original dashboard to reports dashboard for consistency
[#OCD-4956]
