--- src/app/pages/compliance-dashboard/compliance-dashboard-wrapper.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-
-import ChplComplianceDashboard from './compliance-dashboard';
-
-import AppWrapper from 'app-wrapper';
-
-function ChplComplianceDashboardWrapper() {
-  return (
-    <AppWrapper>
-      <ChplComplianceDashboard />
-    </AppWrapper>
-  );
-}
-
-export default ChplComplianceDashboardWrapper;
