+++ src/app/pages/resources/reports-dashboard/reports-dashboard-wrapper.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+import ChplReportsDashboard from './reports-dashboard';
+
+import AppWrapper from 'app-wrapper';
+
+function ChplReportsDashboardWrapper() {
+  return (
+    <AppWrapper>
+      <ChplReportsDashboard />
+    </AppWrapper>
+  );
+}
+
+export default ChplReportsDashboardWrapper;
+
+ChplReportsDashboardWrapper.propTypes = {
+};
