+++ src/app/pages/resources/reports-dashboard/reports-dashboard.jsx
@@ -1,0 +1,168 @@
+import React, { useState, useEffect } from 'react';
+import {
+  Box,
+  Card,
+  CardContent,
+  Typography,
+  Container,
+  Button,
+  makeStyles,
+} from '@material-ui/core';
+
+import { useFetchReportGroupMetadata } from 'api/reports';
+import { eventTrack } from 'services/analytics.service';
+import { useAnalyticsContext } from 'shared/contexts';
+import { palette, theme } from 'themes';
+
+const useStyles = makeStyles({
+  container: {
+    padding: theme.spacing(8),
+    backgroundColor: palette.greyLight,
+    minHeight: 'calc(100vh - 238px)',
+    alignItems: 'flex-start',
+  },
+  stickyCard: {
+    position: 'sticky',
+    top: '116px',
+  },
+  card: {
+    width: '48%',
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+    transition: 'transform 0.3s ease, all 0.2s ease-in-out',
+    cursor: 'pointer',
+    '&:hover': {
+      transform: 'scale(1.02)',
+    },
+  },
+  cardContent: {
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'flex-start',
+    justifyContent: 'center',
+    padding: '32px !important',
+    gap: theme.spacing(1),
+  },
+  cardButtons: {
+    display: 'flex',
+    flexDirection: 'column',
+  },
+});
+
+function ChplReportsDashboard() {
+  const classes = useStyles();
+  const { analytics } = useAnalyticsContext();
+  const [activeReport, setActiveReport] = useState(undefined);
+  const [reportMetadata, setReportMetadata] = useState([]);
+  const { data, isLoading, isSuccess } = useFetchReportGroupMetadata('dashboard');
+
+  useEffect(() => {
+    if (isLoading || !isSuccess) { return; }
+    setReportMetadata(data);
+  }, [data, isLoading, isSuccess]);
+
+  const handleReportChange = (reportKey) => {
+    setActiveReport(reportMetadata.find((metadata) => metadata.reportKey === reportKey));
+    eventTrack({
+      ...analytics,
+      category: 'Dashboard',
+      event: `Navigate to ${reportMetadata.find((metadata) => metadata.reportKey === reportKey)?.title || 'Dashboard'}`,
+    });
+  };
+
+  return (
+    <>
+      <Box bgcolor={palette.white} p={8}>
+        <Container maxWidth="lg">
+          <Typography variant="h1">Dashboard</Typography>
+        </Container>
+      </Box>
+      <Box className={classes.container}>
+        <Container maxWidth="lg">
+          <Box display="flex" alignItems="flex-start" flexDirection="row" gridGap={32} width="100%">
+            <Box maxWidth="350px">
+              <Card className={classes.stickyCard}>
+                <CardContent>
+                  <Box className={classes.cardButtons}>
+                    <Button
+                      style={{ justifyContent: 'flex-start' }}
+                      color="primary"
+                      className={activeReport === undefined ? `${classes.menuButton} active` : classes.menuButton}
+                      onClick={() => handleReportChange(undefined)}
+                    >
+                      Dashboard
+                    </Button>
+                    { reportMetadata
+                      .sort((a, b) => (a.title < b.title ? -1 : 1))
+                      .map((report) => (
+                        <Button
+                          key={`${report.reportKey}-button`}
+                          style={{ justifyContent: 'flex-start' }}
+                          color="primary"
+                          className={`${classes.menuButton} ${activeReport === report.reportKey ? 'active' : ''}`}
+                          onClick={() => handleReportChange(report.reportKey)}
+                          id={`report-${report.reportKey}`}
+                          fullWidth
+                          variant="text"
+                        >
+                          { report.title }
+                        </Button>
+                      ))}
+                  </Box>
+                </CardContent>
+              </Card>
+            </Box>
+            <Box width="100%">
+              { !activeReport && (
+                <Card>
+                  <CardContent>
+                    <Typography gutterBottom variant="h6">
+                      <b>Welcome to the CHPL Dashboard</b>
+                    </Typography>
+                    <Typography gutterBottom>
+                      A dynamic reporting suite powered by PowerBI, providing detailed insights and analytics derived from CHPL data. This tool offers interactive reports with robust click-through capabilities, allowing users to explore and analyze data seamlessly. Each report is designed to be user-friendly, enabling in-depth exploration of key metrics and trends, with the flexibility to dive deeper into the numbers that matter most.
+                    </Typography>
+                    <Box mt={8} mb={4} display="flex" flexDirection="row" flexWrap="wrap" gridGap={32}>
+                      {reportMetadata && reportMetadata.map((report) => (
+                        <Card
+                          key={report.reportKey}
+                          className={classes.card}
+                          onClick={() => handleReportChange(report.reportKey)}
+                        >
+                          <CardContent className={classes.cardContent}>
+                            {report.icon}
+                            <Typography>{ report.title }</Typography>
+                          </CardContent>
+                        </Card>
+                      ))}
+                    </Box>
+                  </CardContent>
+                </Card>
+              )}
+              { activeReport && (
+                <Card
+                  style={{ width: '100%' }}
+                  key={activeReport.reportKey}
+                >
+                  <CardContent>
+                    <iframe
+                      title={activeReport.title}
+                      width="100%"
+                      height={activeReport.height}
+                      src={activeReport.url}
+                      frameBorder="0"
+                      allowFullScreen
+                    />
+                  </CardContent>
+                </Card>
+              )}
+            </Box>
+          </Box>
+        </Container>
+      </Box>
+    </>
+  );
+}
+
+export default ChplReportsDashboard;
