+++ src/app/pages/resources/resources.module.js
@@ -1,8 +1,8 @@
 import { reactToAngularComponent } from '../../services/angular-react-helper';
 
 import ChplCmsLookupWrapper from './cms-lookup/cms-lookup-wrapper';
-import ChplComplianceDashboardWrapper from '../compliance-dashboard/compliance-dashboard-wrapper';
 import ChplDashboardWrapper from './dashboard/dashboard-wrapper';
+import ChplReportsDashboardWrapper from './reports-dashboard/reports-dashboard-wrapper';
 import ChplForgotPassword from './forgot-password/forgot-password-wrapper';
 import ChplNotFound from './not-found/not-found';
 import ChplResourcesApi from './api/api-wrapper';
@@ -17,8 +17,8 @@ angular
     'ngStorage',
   ])
   .component('chplCmsLookupWrapperBridge', reactToAngularComponent(ChplCmsLookupWrapper))
-  .component('chplComplianceDashboardWrapperBridge', reactToAngularComponent(ChplComplianceDashboardWrapper))
   .component('chplDashboardWrapperBridge', reactToAngularComponent(ChplDashboardWrapper))
+  .component('chplReportsDashboardWrapperBridge', reactToAngularComponent(ChplReportsDashboardWrapper))
   .component('chplForgotPasswordBridge', reactToAngularComponent(ChplForgotPassword))
   .component('chplNotFoundBridge', reactToAngularComponent(ChplNotFound))
   .component('chplResourcesApiBridge', reactToAngularComponent(ChplResourcesApi))
